requirements.txt
# 核心框架（版本严格匹配原项目）
pydantic==1.10
langsmith>=0.1.25,<0.2
langchain-community>=0.0.27,<0.1
langchain-openai>=0.0.8,<0.1
langchain-google-genai>=0.0.6,<0.1
langchain-anthropic>=0.1.4,<0.2
langchain-fireworks>=0.1.1,<0.2
langserve[server]>=0.0.39,<0.1

# API 服务
fastapi>=0.103.1,<0.104
uvicorn>=0.23.2,<0.24

# 数据库与存储
psycopg2-binary>=2.9.7,<3.0
weaviate-client>=3.26.7,<4.0

# 工具
beautifulsoup4>=4.12.2,<5.0
lxml>=4.9.3,<5.0
pillow>=10.2.0,<11.0
tiktoken>=0.5.2,<0.6
voyageai>=0.1.4,<0.2
cohere==4.45
auth>=0.5.3,<0.6
python-dotenv

# 开发（可选）
ruff>=0.2.2,<0.3